Restore validates_present Boolean auto-validation.
dkubb requested this change based on my previous commit. He said
"I'd prefer to keep as many auto validations in place as possible (and
add more if we missed anything) than to remove them."  The discussion
is summarized on the github discussion for commit ba2e873bba.

We agreed that a Boolean property should auto-generate the
validates_present validation.  A boolean is considered present if it is
true or false, but not nil.  Note that the behavior for boolean is
different than for other property types, which use #blank? instead.