[merb_dm] Connection info now has its own key
Connection info is now set in its own key within the :merb_datamapper
hash.  This change will probably be unnoticed by all merb_datamapper
users, as currently the only way this will be triggered is to have a
config/database.yml file -- which is intended to be ignored.

It might even be removed in subsequent revisions.