[merb_dm] new Merb sessions.
This commit fixes merb_datamapper for the new session implementation,
introduced in Merb 0.9.6.  The merb version required has been bumped
appropriately.

In addition, as part of the updating to the new version,
merb_datamapper is now initialized via a proper Merb BootLoader, which
has the side effect of allowing DataMapper to integrate with the Merb
logger, at least if data-store connections are setup via the
database.yml file.  The default log setting of :debug in development
means you'll see the statements DataMapper generates as it queries
your data.