[merb_dm] stepping down the dependencies
merb_datamapper is using no private DataMapper behavior which would require
a strict version locking with dm-core, and given the differing merb and
datamapper release schedules, it seems easier this way.  It will require a
little more vigilance in the future, should this state of affairs change, but
for now it seems correct.