fixes to simplify STI and avoid namespace clashes
the auto-generated type value is now called
couchdb_type, as this is less likely to be used
by accident (the adapter overwrites this value if
it happens to exist, so it won't work if you use
it).  Also use property.typecast, otherwise the
discriminator values are still strings instead
of classes and things die.