[dm-couchdb-adapter] #couchdb_types_condition doesn't return subclasses during #auto_migrate!
The method is supposed to return a javascript query string that
selects the parent model and all its subclasses. Works fine in the
console but fails to return the expected subclasses during migration,
likely because couch models include DM::CouchResource instead of
DM::Resource. Added failing specs to demonstrate the problem.