DataMapper::Validate is now DataMapper::Validations
Since this change has the potential to completely
break community plugins that rely on doing stuff
like DataMapper.const_defined?('Validate'), we
still provide a constant alias for the time being.

  DataMapper::Validate

will continue to work for some time but will not
be supported once we hit 1.0. The release notes
for the next release will include a warning that
this namespace is deprecated and

  DataMapper::Validations

should be used instead. In fact, we might add API
to dm-validations that will allow to check for its
presence without the need of doing const_defined?
checks.

Big thanks to hipe on IRC, for totally helping me
out with the search/replace regex that made this
refactoring bearable.

  search:  \b(Validate)($|:)
  replace: Validations$2

While it didn't catch *all* occurrences, it only
left out 3 places in the whole of dm-more. Note
that it didn't touch comments that started with

  # Validate something for some reason